% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackageWilbourne-package.R
\docType{package}
\name{testpackageWilbourne-package}
\alias{testpackageWilbourne}
\alias{testpackageWilbourne-package}
\title{testpackageWilbourne: Demos Creating a Package}
\description{
What the package does (not very much).
}
\author{
\strong{Maintainer}: David Wilbourne \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
